+++ internal/topic/topicwriterinternal/writer_reconnector.go
@@ -573,12 +573,12 @@ func createRawMessageData(
 	res.SeqNo = mess.SeqNo
 
 	switch {
-	case mess.Partitioning.hasPartitionID:
+	case mess.futurePartitioning.hasPartitionID:
 		res.Partitioning.Type = rawtopicwriter.PartitioningPartitionID
-		res.Partitioning.PartitionID = mess.Partitioning.partitionID
-	case mess.Partitioning.messageGroupID != "":
+		res.Partitioning.PartitionID = mess.futurePartitioning.partitionID
+	case mess.futurePartitioning.messageGroupID != "":
 		res.Partitioning.Type = rawtopicwriter.PartitioningMessageGroupID
-		res.Partitioning.MessageGroupID = mess.Partitioning.messageGroupID
+		res.Partitioning.MessageGroupID = mess.futurePartitioning.messageGroupID
 	default:
 		// pass
 	}